added working version of the project
