+++ include/geometry/sphere.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "../raytracer/common_libs.h"
+#include "vector.h"
+
+class Sphere {
+public:
+    Sphere(Vector center, double radius)
+        : center_(center)
+        , radius_(radius) {}
+
+    const Vector& GetCenter() const {
+        return center_;
+    }
+
+    double GetRadius() const {
+        return radius_;
+    }
+
+private:
+    Vector center_;
+    double radius_;
+};
