+++ include/raytracer/render_options.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum class RenderMode {
+    kDepth, kNormal, kFull
+};
+
+struct RenderOptions {
+    int depth;
+    RenderMode mode;
+};
