--- raytracer/common_libs.h
@@ -1,22 +1,0 @@
-#pragma once
-
-#define _USE_MATH_DEFINES
-#include <math.h>
-#include <random>
-#include <queue>
-#include <utility>
-#include <limits>
-#include <optional>
-#include <initializer_list>
-#include <cassert>
-#include <array>
-#include <cmath>
-#include <iostream>
-#include <algorithm>
-#include <memory>
-#include <string_view>
-#include <vector>
-#include <map>
-#include <string>
-#include <fstream>
-
